fixed doom1 timedemo 2. bug caused by huge sector with 374 lines - guess linecount 8 bits wasnt enough. made linecount 16 bits again, this made sectors 17 bytes each so i moved lightlevels out into their own field to make sectors 16 bytes again, but while i thought that would lead to more optimized (?) code apparently watcom doesnt do anything with that, may go back to 17 byte structs?

+++ p_floor.c
@@ -523,7 +523,7 @@ EV_BuildStairs
 	short_height_t sectorfloorheight;
 	uint8_t sectorfloorpic;
 	int16_t sectorlinesoffset;
-	uint8_t sectorlinecount;
+	int16_t sectorlinecount;
 	int16_t secnumlist[MAX_ADJOINING_SECTORS];
 	int16_t		j = 0;
 	sector_t __far* sector;
